moves configure_monitoring.go to own package 'monitoring'
also moves helper.go from 'event_handler' to own package 'event'
